+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Current (in progress)
 
 - Fix `<read-more>` component height when it contains `<img>` [65](https://github.com/etalab/udata-front/pull/65)
+- Update reuse style [#52](https://github.com/etalab/udata-front/pull/52)
 
 ## 1.2.3 (2022-01-27)